refactor: render home page with latest documents page
