+++ src/views/Documents/index.ts
@@ -2,6 +2,12 @@
 import List from './list.view'
 import Create from './create.view'
 import Show from './show.view'
+import Latest from './latest.view'
 
 // exports
-export { List as ListDocuments, Create as CreateDocument, Show as ShowDocument }
+export {
+	Latest as LatestDocuments,
+	List as ListDocuments,
+	Create as CreateDocument,
+	Show as ShowDocument,
+}
